refactor: refactor build gradle & convention plugins
