+++ build-logic/convention/build.gradle.kts
@@ -2,7 +2,6 @@ plugins {
     `kotlin-dsl`
 
     alias(libs.plugins.jetbrains.kotlin.jvm)
-    alias(libs.plugins.kotlin.serialization)
 }
 
 group = "com.emotionstorage.buildlogic"
@@ -60,9 +59,9 @@ gradlePlugin {
         /**
          * java kotlin library plugins
          */
-        create("javaKotlinLibrary"){
+        create("kotlinLibrary"){
             id = "com.emotionstorage.convention.kotlin.library"
-            implementationClass = "com.emotionstorage.convention.JavaKotlinLibraryConventionPlugin"
+            implementationClass = "com.emotionstorage.convention.KotlinLibraryConventionPlugin"
         }
         create("hilt"){
             id = "com.emotionstorage.convention.kotlin.library.hilt"
