+++ build-logic/convention/src/main/java/com/emotionstorage/convention/AndroidApplicationHiltConventionPlugin.kt
@@ -1,11 +1,9 @@
 package com.emotionstorage.convention
 
-import com.android.build.api.dsl.ApplicationExtension
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalogsExtension
 import org.gradle.kotlin.dsl.getByType
-import org.gradle.kotlin.dsl.findByType
 import org.gradle.kotlin.dsl.dependencies
 
 class AndroidApplicationHiltConventionPlugin : Plugin<Project> {
@@ -16,17 +14,12 @@ class AndroidApplicationHiltConventionPlugin : Plugin<Project> {
             with(pluginManager) {
                 apply("com.google.dagger.hilt.android")
                 apply("com.google.devtools.ksp")
+            }
 
-                extensions.findByType<ApplicationExtension>()?.let {
-                    val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
-
-                    it.apply {
-                        dependencies {
-                            add("implementation", libs.findBundle("hilt-android").get())
-                            add("ksp", libs.findLibrary("google-dagger-hilt-compiler").get())
-                        }
-                    }
-                }
+            val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
+            dependencies {
+                add("implementation", libs.findBundle("hilt-android").get())
+                add("ksp", libs.findLibrary("google-dagger-hilt-compiler").get())
             }
         }
     }
