+++ build-logic/convention/src/main/java/com/emotionstorage/convention/KotlinLibraryConventionPlugin.kt
@@ -1,0 +1,41 @@
+package com.emotionstorage.convention
+
+import com.android.build.api.dsl.CommonExtension
+import org.gradle.api.JavaVersion
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.api.artifacts.VersionCatalogsExtension
+import org.gradle.kotlin.dsl.getByType
+import org.gradle.kotlin.dsl.withType
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+import org.gradle.api.plugins.JavaPluginExtension
+import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.dependencies
+
+class KotlinLibraryConventionPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        println("*** KotlinLibraryConventionPlugin invoked ***")
+
+        with(target) {
+            with(pluginManager) {
+                apply("java-library")
+                apply("org.jetbrains.kotlin.jvm")
+            }
+
+            extensions.getByType<JavaPluginExtension>().apply {
+                sourceCompatibility = JavaVersion.VERSION_17
+                targetCompatibility = JavaVersion.VERSION_17
+            }
+
+            tasks.withType<KotlinCompile>().configureEach {
+                @Suppress("DEPRECATION")
+                kotlinOptions.jvmTarget = JavaVersion.VERSION_17.toString()
+            }
+
+            val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
+            dependencies {
+                add("implementation", libs.findLibrary("kotlinx.serialization.json").get())
+            }
+        }
+    }
+}
