+++ gradle/libs.versions.toml
@@ -27,6 +27,7 @@ serializationConverter = "1.0.0"
 
 ksp = "2.0.21-1.0.27"
 dependencyGraphGenerator = "0.7.0"
+googleServices = "4.4.3"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -66,7 +67,6 @@ squareup-retrofit = { group = "com.squareup.retrofit2", name = "retrofit", versi
 kotlinx-serialization-json = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version.ref = "serialization" }
 kotlinx-serialization-converter = { group = "com.jakewharton.retrofit", name = "retrofit2-kotlinx-serialization-converter", version.ref = "serializationConverter" }
 
-
 android-gradlePlugin = { group = "com.android.tools.build", name = "gradle", version.ref = "agp" }
 kotlin-gradlePlugin = { group = "org.jetbrains.kotlin", name = "kotlin-gradle-plugin", version.ref = "kotlin" }
 
@@ -85,8 +85,8 @@ compose = [
     "androidx-compose-runtime",
     "androidx-compose-material3-android",
     "androidx-navigation-compose",
-#    "coil-compose",
-#    "coil-network-ktor"
+    #    "coil-compose",
+    #    "coil-network-ktor"
 ]
 hilt-android = [
     "hilt-navigation-compose",
@@ -109,7 +109,6 @@ retrofit = [
     "okhttp3-urlconnection",
     "okhttp3-logging-interceptor",
     "squareup-retrofit",
-    "kotlinx-serialization-json",
     "kotlinx-serialization-converter"
 ]
 compose-debug = [
@@ -138,6 +137,6 @@ jetbrains-kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "jetbrai
 google-dagger-hilt-android = { id = "com.google.dagger.hilt.android", version.ref = "daggerHilt" }
 google-devtools-ksp = { id = "com.google.devtools.ksp", version.ref = "ksp" }
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
+google-services = { id = "com.google.gms.google-services", version.ref = "googleServices" }
 vanniktech-dependency-graph-generator = { id = "com.vanniktech.dependency.graph.generator", version.ref = "dependencyGraphGenerator" }
-
